rpi/requirements.txt
rpi-ws281x==5.0.0
requests==2.31.0
python-dotenv==1.0.0
